# Optimize values() calls