--- a/fiftyone/core/collections.py
+++ b/fiftyone/core/collections.py
@@ -10630,6 +10630,9 @@ class SampleCollection(object):
             if _group_slices:
                 group_slices.update(_group_slices)
 
+            # Optimization: some aggregations are allowed to change field name
+            big_field = getattr(aggregation, "_big_field", big_field)
+
             try:
                 assert len(_pipeline) == 1
                 project[big_field] = _pipeline[0]["$project"][big_field]