## Change log

This PR experiments with optimizing `values("id")` calls.

Currently, `values("id")` executes the following pipeline:

```py
pipeline = [
    {
        "$project": {
            "value": {
                "$cond": {
                    "if": {"$gt": ["$_id", None]},
                    "then": {"$toString": "$_id"},
                    "else": None,
                },
            },
        },
    },
]
```

I wanted to see how much faster `values("id")` would be if the following pipeline were instead used:

```py
pipeline = [{"$project": {"_id": 1}}]
```

with the `str()` conversion instead happening in Python rather than in the database.

Additionally, `values("id")` does not leverage the `_id` field index in either of the above pipelines by default. It uses a `COLLSCAN`. So, I also wanted to see how much faster/slower both pipelines are when they use an `IXSCAN`, which can be forced via `hint={"_id": 1}`.

## Findings 

The benchmarking section below shows the following findings on a local database with a 500k sample dataset:
- The `{"$project": {"_id": 1}}` pipeline runs in 2.1 seconds compared to the original pipeline, which runs in 3.4 seconds
- Interestingly, adding `hint={"_id": 1}` makes both pipelines 0.4-0.6 seconds *slower* 

## Create large dataset

```py
import numpy as np

import fiftyone as fo
import fiftyone.zoo as foz

def expand_dataset(dataset, target_num_samples):
    num_doubles = int(np.ceil(np.log2(target_num_samples / len(dataset))))
    with fo.ProgressBar(start_msg="Expanding dataset") as pb:
        for i in pb(list(range(1, num_doubles + 1))):
            # Cloning avoids the halloween problem, which I observed if directly using `dataset` in-place of `tmp` here
            # https://www.mongodb.com/docs/manual/reference/operator/aggregation/merge/#output-to-the-same-collection-that-is-being-aggregated
            if i == num_doubles:
                tmp = dataset[:(target_num_samples - len(dataset))].clone()
            else:
                tmp = dataset.clone()

            dataset.add_collection(tmp, new_ids=True)
            tmp.delete()

dataset = foz.load_zoo_dataset(
    "cifar10",
    split="train",
    dataset_name="zzz",
    persistent=True,
)

expand_dataset(dataset, 500000)
```

## Test real usage

```py
import fiftyone as fo
import eta.core.utils as etau

dataset = fo.load_dataset("zzz")
dataset.clone_sample_field("id", "id2")
# dataset.create_index("id2")

# Uses optimized values() call
# Time elapsed: 2.7 seconds
with etau.Timer():
    _ = dataset.values("id")

# Using existing values() call
# Time elapsed: 3.8 seconds
with etau.Timer():
    _ = dataset.values("id2")
```

## Benchmarking

```py
import fiftyone as fo
import eta.core.utils as etau

dataset = fo.load_dataset("zzz")

# Current behavior: in-database string conversion
# Time elapsed: 2.8 seconds
with etau.Timer():
    pipeline = [
        {
            "$project": {
                "value": {
                    "$cond": {
                        "if": {"$gt": ["$_id", None]},
                        "then": {"$toString": "$_id"},
                        "else": None,
                    },
                },
            },
        },
    ]
    results = dataset._sample_collection.aggregate(pipeline)
    ids = [d["value"] for d in results]

# With field inclusion
# Time elapsed: 2.1 seconds
with etau.Timer():
    pipeline = [{"$project": {"_id": 1}}]
    results = dataset._sample_collection.aggregate(pipeline)
    ids = [d["_id"] for d in results]
    ids = [str(_id) for _id in ids]

# With field inclusion + index hint
# Time elapsed: 2.5 seconds
with etau.Timer():
    pipeline = [{"$project": {"_id": 1}}]
    results = dataset._sample_collection.aggregate(pipeline, hint={"_id": 1})
    ids = [d["_id"] for d in results]
    ids = [str(_id) for _id in ids]

# With field setting + hint
# Time elapsed: 3.4 seconds
with etau.Timer():
    pipeline = [{"$project": {"value": "$_id"}}]
    results = dataset._sample_collection.aggregate(pipeline, hint={"_id": 1})
    ids = [d["value"] for d in results]
    ids = [str(_id) for _id in ids]
```