@kaixi-wang 2025-02-16
Interesting... the project _id and hint at the db level I would expect to be doing the same thing...

When I run it, even on a small dataset, hint is always faster...
<img width="596" alt="image" src="https://github.com/user-attachments/assets/627357c3-fbfd-4698-b5e5-938fa1f08d78" />

I'll check the explain plans/do a little more investigating