@kaixi-wang 2025-02-16
Also the results on 11million samples: 
<img width="621" alt="image" src="https://github.com/user-attachments/assets/568d39d6-f621-4ad2-9de7-601765ff2479" />